create core.rs && split code
